Fixe bug of incresing nyan speed if a key is holded.
Also moved print function calls from read_input to game_handler. This might
cause troubles if we want to detect collisions between nyan and the items in
the world, but it seem to be the cleanes way form implementation. Read_input
is only used to read user input and to move objects or detect collisions.
After this the objects will be printed to screen in game_handler.