Allow to initialize the timer with fps value.
For now this isn't required, but later we can speed the whole game up by
reinitializing the timer. Maybe there should be another function in main.c
that triggers the timer setting. I think it wouldn't be a good choice to make
the set_timer() function none static for this case.