Implemented array based pflatform generation (#3).
This allows to get rid of the position of the platform to check if nyan has
ground under the feets. I know that the random stuff isn't pretty nice,
sometime there are to large gaps between the platforms and wenn need more CPU
resources for the randomizing. I think there are a lot of improvements
possible, but this is only a first step to implement the jumping of nyan.