Fixed none terminated special modes after pause (#17).
If nyan entered a special cat mode or collected a gem, a event to remove that
mode was added to the event queue. If the game where pauses within such an
special mode, this specialmode would never been reset after pause is released
because the queue must be cleared before another mode is entered.

Now we backup the relevant event data for nyan and restore them on start of
game or after the pause.